requirements.txt
# requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
alembic==1.10.4
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.0.1
openai==0.27.6
google-cloud-vision==3.4.0
pdf2image==1.16.3
markitdown==0.1.0
playwright==1.32.1
python-dotenv==1.0.0
holidays==0.25
celery==5.2.7
redis==4.5.5
pytest==7.3.1
httpx==0.24.0